@@ -1,12 +1,12 @@
 {
   "name": "loader-utils",
-  "version": "1.2.3",
+  "version": "2.0.2",
   "author": "Tobias Koppers @sokra",
   "description": "utils for webpack loaders",
   "dependencies": {
     "big.js": "^5.2.2",
-    "emojis-list": "^2.0.0",
-    "json5": "^1.0.1"
+    "emojis-list": "^3.0.0",
+    "json5": "^2.1.2"
   },
   "scripts": {
     "lint": "eslint lib test",
@@ -21,16 +21,16 @@
     "url": "https://github.com/webpack/loader-utils.git"
   },
   "engines": {
-    "node": ">=4.0.0"
+    "node": ">=8.9.0"
   },
   "devDependencies": {
-    "coveralls": "^3.0.2",
-    "eslint": "^5.11.0",
-    "eslint-plugin-node": "^8.0.0",
-    "eslint-plugin-prettier": "^3.0.0",
-    "jest": "^21.2.1",
-    "prettier": "^1.15.3",
-    "standard-version": "^4.0.0"
+    "coveralls": "^3.0.9",
+    "eslint": "^6.8.0",
+    "eslint-plugin-node": "^11.0.0",
+    "eslint-plugin-prettier": "^3.1.2",
+    "jest": "^25.1.0",
+    "prettier": "^1.19.1",
+    "standard-version": "^7.1.0"
   },
   "main": "lib/index.js",
   "files": [
